refactor state helpers
